Fix compilation warnings on Win64
Fix compilation warnings on Win64, as shown below, by using proper long types
```
  gcompletion.c(47): warning C4028: formal parameter 3 different from declaration [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gconvert.c(105): warning C4133: 'function': incompatible types - from 'gsize *' to 'size_t *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gconvert.c(460): warning C4133: 'function': incompatible types - from 'size_t *' to 'gsize *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gdataset.c(552): warning C4311: 'type cast': pointer truncation from 'gpointer' to 'guint' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  giochannel.c(1238): warning C4133: 'function': incompatible types - from 'size_t *' to 'gsize *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gmem.c(662): warning C4311: 'type cast': pointer truncation from 'gpointer' to 'guint' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  ...
```

+++ graf2d/win32gdk/gdk/src/glib/glibconfig-bogus.h
@@ -84,11 +84,11 @@ typedef unsigned long long guint64;
 #define G_GINT64_FORMAT "I64i"
 #define G_GUINT64_FORMAT "I64u"
 
-typedef gint32 gssize;
-typedef guint32 gsize;
+typedef intptr_t  gssize;
+typedef uintptr_t gsize;
 
-#define GPOINTER_TO_INT(p)	((gint)(p))
-#define GPOINTER_TO_UINT(p)	((guint)(p))
+#define GPOINTER_TO_INT(p)	((intptr_t)   (p))
+#define GPOINTER_TO_UINT(p)	((uintptr_t)  (p))
 
 #define GINT_TO_POINTER(i)	((gpointer)(i))
 #define GUINT_TO_POINTER(u)	((gpointer)(u))
